MergedTreeBuilder: move to new module
`MergedTreeBuilder` doesn't depend on any internals of `MergedTree`,
and I'm hoping to remove the existing `TreeBuilder` and only have
`MergedTreeBuilder`, which will make `MergedTreeBuilder` larger. Let's
move it out before that.
