+++ lib/src/fix.rs
@@ -26,8 +26,8 @@ use jj_lib::backend::CommitId;
 use jj_lib::backend::FileId;
 use jj_lib::backend::TreeValue;
 use jj_lib::matchers::Matcher;
-use jj_lib::merged_tree::MergedTreeBuilder;
 use jj_lib::merged_tree::TreeDiffEntry;
+use jj_lib::merged_tree_builder::MergedTreeBuilder;
 use jj_lib::repo::MutableRepo;
 use jj_lib::repo::Repo as _;
 use jj_lib::repo_path::RepoPathBuf;
