+++ lib/src/lib.rs
@@ -73,6 +73,7 @@ pub mod lock;
 pub mod matchers;
 pub mod merge;
 pub mod merged_tree;
+pub mod merged_tree_builder;
 pub mod object_id;
 pub mod op_heads_store;
 pub mod op_store;
