+++ lib/src/local_working_copy.rs
@@ -103,8 +103,8 @@ use crate::merge::MergeBuilder;
 use crate::merge::MergedTreeValue;
 use crate::merge::SameChange;
 use crate::merged_tree::MergedTree;
-use crate::merged_tree::MergedTreeBuilder;
 use crate::merged_tree::TreeDiffEntry;
+use crate::merged_tree_builder::MergedTreeBuilder;
 use crate::object_id::ObjectId as _;
 use crate::op_store::OperationId;
 use crate::ref_name::WorkspaceName;
