+++ lib/src/rewrite.rs
@@ -44,8 +44,8 @@ use crate::matchers::Visit;
 use crate::merge::Diff;
 use crate::merge::Merge;
 use crate::merged_tree::MergedTree;
-use crate::merged_tree::MergedTreeBuilder;
 use crate::merged_tree::TreeDiffEntry;
+use crate::merged_tree_builder::MergedTreeBuilder;
 use crate::repo::MutableRepo;
 use crate::repo::Repo;
 use crate::repo_path::RepoPath;
